+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 # ServiceStackIDEA Changelog
 
 ## [Unreleased]
+- Fix WebStorm Add ServiceStack Reference
+- Fix Java/Kotlin Add ServiceStack Reference for Intellij + Android Studio
+- Upgrade CI process
+- Restrict support to 192+
 ## [1.1.2-beta]
 ### Added
 - Fix WebStorm Add ServiceStack Reference